Changed distmod structure
